+++ conversion/src/sumo/nodes.rs
@@ -11,7 +11,7 @@ pub struct Node {
     #[serde(rename = "@id")]
     pub id: String,
     #[serde(rename = "@x")]
-    pub x: f32,
+    pub x: f64,
     #[serde(rename = "@y")]
-    pub y: f32,
+    pub y: f64,
 }
